requirements.txt
fastapi
uvicorn
sqlalchemy
pyjwt
passlib[bcrypt]
pydantic[email]
python-multipart
types-passlib
ruff
pytz